refactor: moved contributor model types to types directory
